Move upload and git-push commands
