+++ cli/cmd/org/switch.go
@@ -1,6 +1,7 @@
 package org
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/rilldata/rill/cli/pkg/cmdutil"
@@ -74,4 +75,35 @@ func SwitchSelectFlow(orgs []*adminv1.Organization) (string, error) {
 	}
 
 	return cmdutil.SelectPrompt("Select default org.", orgNames, org)
+}
+
+// SetDefaultOrg sets a default org for the user if user is part of any org.
+func SetDefaultOrg(ctx context.Context, ch *cmdutil.Helper) error {
+	c, err := ch.Client()
+	if err != nil {
+		return err
+	}
+
+	res, err := c.ListOrganizations(ctx, &adminv1.ListOrganizationsRequest{})
+	if err != nil {
+		return fmt.Errorf("listing orgs failed: %w", err)
+	}
+
+	if len(res.Organizations) == 1 {
+		ch.Org = res.Organizations[0].Name
+		if err := dotrill.SetDefaultOrg(ch.Org); err != nil {
+			return err
+		}
+	} else if len(res.Organizations) > 1 {
+		orgName, err := SwitchSelectFlow(res.Organizations)
+		if err != nil {
+			return fmt.Errorf("org selection failed %w", err)
+		}
+
+		ch.Org = orgName
+		if err := dotrill.SetDefaultOrg(ch.Org); err != nil {
+			return err
+		}
+	}
+	return nil
 }
